let more IPs in
previously always using MAAS to do final deploy; which was ending up with nodes where the last octet of the IP was three digits. Now using a coreos-install script to write a disk image prior to MAAS's deploy stage which leaves hosts with IPs where the last octet is two digits. So this change lets more IPs in